feat(model_router.go): add SelectModelRequest and SelectModelResponse types to support model selection functionality refactor(select_model.go): remove redundant SelectModelRequest and SelectModelResponse types as they are now defined in model_router.go
